Chore: 마이페이지와 Setting 분리 및 수정디자인 반영
