+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover1.imageset/Contents.json
