+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover1.imageset/LeftTop@1x.png
