+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover2.imageset/RightTop@3x.png
