+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover3.imageset/RightBottom@1x.png
