+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover3.imageset/RightBottom@2x.png
