+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover4.imageset/LeftBottom@1x.png
