+++ OneByte/Resource/Assets.xcassets/TabClover/ColorClover4.imageset/LeftBottom@2x.png
