+++ OneByte/Resource/Assets.xcassets/TabClover/Contents.json
