+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover1.imageset/Contents.json
