+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover1.imageset/LeftTop2@3x.png
