+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover2.imageset/RightTop2@2x.png
