+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover3.imageset/RightBottom2@2x.png
