+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover4.imageset/LeftBottom2@1x.png
