+++ OneByte/Resource/Assets.xcassets/TabClover/GrayClover4.imageset/LeftBottom2@2x.png
